fix Apache Restart during Update (cert creation)

+++ sbin/makewwwcert.sh
@@ -8,5 +8,10 @@ mv newcert.pem $CATOP/certs/wwwcert.pem
 mv newkeywp.pem $CATOP/private/wwwkeywp.pem
 mv newkey.pem $CATOP/private/wwwkey.pem
 rm newreq.pem
-systemctl stop apache2
-systemctl start apache2
+systemctl is-enabled apache2
+if [ $? -eq 0 ]
+then
+  echo "restart Apache"
+  systemctl stop apache2
+  systemctl start apache2
+fi
